requirements.txt
# Streamlit
streamlit

# Visualization
plotly
scikit-image
opencv-python

# Data handling
scipy

# Preprocessing
nibabel

# Deep learning
sklearn
tensorflow-cpu

